+++ src/main.cpp
@@ -11,7 +11,7 @@ using namespace keybinds;
 #endif
 
 $execute {
-	std::string dataDir = Mod::get()->getSaveDir().generic_string();
+	std::string dataDir = string::pathToString(Mod::get()->getSaveDir());
 	std::filesystem::create_directories(fmt::format("{}/saves/main/", dataDir));
 	std::filesystem::create_directories(
 		fmt::format("{}/saves/editor/", dataDir)